backend-dev into develop

See merge request s12-final/S12P31B103!7

+++ server/src/main/java/com/ssafy/odab/domain/learning/repository/LastLearningDateRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.odab.domain.learning.repository;
 import com.ssafy.odab.domain.learning.entity.LastLearningDate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LastLearningDateRepository extends JpaRepository<LastLearningDate, Long> {
+public interface LastLearningDateRepository extends JpaRepository<LastLearningDate, Integer> {
 
 }
